requirements.txt
backports.os==0.1.1
click==8.0.1
colorama==0.4.4
Flask==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.1
lxml==4.6.3
MarkupSafe==2.0.1
Pillow==8.3.1
PyPDF2==1.26.0
python-pptx==0.6.18
Wand==0.5.9
Werkzeug==2.0.1
XlsxWriter==1.4.5